Use long for ids everywhere.

+++ subprocess/src/main/java/net/creeperhost/creeperlauncher/api/data/instances/InstanceInstallModData.java
@@ -8,8 +8,8 @@ import java.util.UUID;
 public class InstanceInstallModData extends BaseData {
 
     public UUID uuid;
-    public int modId;
-    public int versionId;
+    public long modId;
+    public long versionId;
 
     public static class Reply extends BaseData {
